requirements.txt
Django==5.2.4
djangorestframework==3.16.1
django-cors-headers==4.7.0
pymongo==4.13.2
python-decouple==3.8
Pillow==11.3.0
Unidecode==1.4.0
whitenoise==6.9.0
gunicorn==23.0.0
